move AddBountyAmountViewModel to commons
